move code from rc.lua to files
