+++ rules/client.lua
@@ -1,0 +1,8 @@
+--
+-- Created by IntelliJ IDEA.
+-- User: oleg
+-- Date: 28.06.15
+-- Time: 2:47
+-- To change this template use File | Settings | File Templates.
+--
+
